QN 31 & 32
QN 31 returns the number of users with an input username. So this can be
used during account creation. If numUsers is >0 then account can't be
created.
QN 32 returns the number of users with an input username and password.
This can be used to verify that a given username and password are
correct to allow changing of the password.