Added file function-scripts.sql
Added a .sql file for adding the functions to calculate time and
distance differences between previous and next points by a mobile_id. It
also adds functions to calculate average speed based on the time and
spatial differences in the points. The functions are mphtonext,
miledifftonext, seconddifftonext, mphfromprevious, miledifffromprevious,
and seconddifffromprevious.